+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLsurface.cpp
@@ -233,7 +233,7 @@ bool surface_exists(int id)
     return size_t(id) < enigma::surface_max && enigma::surface_array[id] != NULL;
 }
 
-void draw_surface(int id, double x, double y)
+void draw_surface(int id, float x, float y)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -252,7 +252,7 @@ void draw_surface(int id, double x, double y)
   glPopAttrib();
 }
 
-void draw_surface_stretched(int id, double x, double y, double w, double h)
+void draw_surface_stretched(int id, float x, float y, float w, float h)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -269,7 +269,7 @@ void draw_surface_stretched(int id, double x, double y, double w, double h)
   glPopAttrib();
 }
 
-void draw_surface_part(int id,double left,double top,double width,double height,double x,double y)
+void draw_surface_part(int id, float left, float top, float width, float height, float x, float y)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -292,7 +292,7 @@ void draw_surface_part(int id,double left,double top,double width,double height,
   glPopAttrib();
 }
 
-void draw_surface_tiled(int id,double x,double y)
+void draw_surface_tiled(int id, float x, float y)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -323,7 +323,7 @@ void draw_surface_tiled(int id,double x,double y)
   glPopAttrib();
 }
 
-void draw_surface_tiled_area(int id,double x,double y,double x1,double y1,double x2,double y2)
+void draw_surface_tiled_area(int id, float x, float y, float x1, float y1, float x2, float y2)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -373,7 +373,7 @@ void draw_surface_tiled_area(int id,double x,double y,double x1,double y1,double
   glPopAttrib();
 }
 
-void draw_surface_ext(int id,double x,double y,double xscale,double yscale,double rot,int color,double alpha)
+void draw_surface_ext(int id, float x, float y, float xscale, float yscale, double rot, int color, double alpha)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -402,7 +402,7 @@ void draw_surface_ext(int id,double x,double y,double xscale,double yscale,doubl
   glPopAttrib();
 }
 
-void draw_surface_stretched_ext(int id,double x,double y,double w,double h,int color,double alpha)
+void draw_surface_stretched_ext(int id, float x, float y, float w, float h, int color, double alpha)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -423,7 +423,7 @@ void draw_surface_stretched_ext(int id,double x,double y,double w,double h,int c
   glPopAttrib();
 }
 
-void draw_surface_part_ext(int id,double left,double top,double width,double height,double x,double y,double xscale,double yscale,int color,double alpha)
+void draw_surface_part_ext(int id, float left, float top, float width, float height, float x, float y, float xscale, float yscale, int color, double alpha)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -447,7 +447,7 @@ void draw_surface_part_ext(int id,double left,double top,double width,double hei
   glPopAttrib();
 }
 
-void draw_surface_tiled_ext(int id,double x,double y,double xscale,double yscale,int color,double alpha)
+void draw_surface_tiled_ext(int id, float x, float y, float xscale, float yscale, int color, double alpha)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -478,7 +478,7 @@ void draw_surface_tiled_ext(int id,double x,double y,double xscale,double yscale
   glPopAttrib();
 }
 
-void draw_surface_tiled_area_ext(int id,double x,double y,double x1,double y1,double x2,double y2, double xscale, double yscale, int color, double alpha)
+void draw_surface_tiled_area_ext(int id, float x, float y, float x1, float y1, float x2, float y2, float xscale, float yscale, int color, double alpha)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -528,7 +528,7 @@ void draw_surface_tiled_area_ext(int id,double x,double y,double x1,double y1,do
   glPopAttrib();
 }
 
-void draw_surface_general(int id, double left,double top,double width,double height,double x,double y,double xscale,double yscale,double rot,int c1,int c2,int c3,int c4,double a1,double a2,double a3,double a4)
+void draw_surface_general(int id, float left, float top, float width, float height, float x, float y, float xscale, float yscale, double rot, int c1, int c2, int c3, int c4, double a1, double a2, double a3, double a4)
 {
   get_surface(surf,id);
   texture_use(surf->tex);
@@ -678,7 +678,7 @@ int surface_save(int id, string filename)
 	return 1;
 }
 
-int surface_save_part(int id, string filename,unsigned x,unsigned y,unsigned w,unsigned h)
+int surface_save_part(int id, string filename, unsigned x, unsigned y, unsigned w, unsigned h)
 {
     get_surfacev(surf,id,-1);
 	FILE *bmp=fopen(filename.c_str(),"wb");
@@ -731,7 +731,7 @@ int surface_save_part(int id, string filename,unsigned x,unsigned y,unsigned w,u
 	return 1;
 }
 
-int sprite_create_from_surface(int id,int x,int y,int w,int h,bool removeback,bool smooth,int xorig,int yorig)
+int sprite_create_from_surface(int id, int x, int y, int w, int h, bool removeback, bool smooth, int xorig, int yorig)
 {
     get_surfacev(surf,id,-1);
     int full_width=nlpo2dc(w)+1, full_height=nlpo2dc(h)+1;
@@ -751,7 +751,7 @@ int sprite_create_from_surface(int id,int x,int y,int w,int h,bool removeback,bo
     return sprid;
 }
 
-void surface_copy_part(int destination,double x,double y,int source,int xs,int ys,int ws,int hs)
+void surface_copy_part(int destination, float x, float y, int source, int xs, int ys, int ws, int hs)
 {
     get_surface(ssurf,source);
     get_surface(dsurf,destination);
@@ -775,7 +775,7 @@ void surface_copy_part(int destination,double x,double y,int source,int xs,int y
 	delete[] surfbuf;
 }
 
-void surface_copy(int destination,double x,double y,int source)
+void surface_copy(int destination, float x, float y, int source)
 {
     get_surface(ssurf,source);
     get_surface(dsurf,destination);
